# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": " This API Provides the fundamentals of the back-end system of the **Ideal Place To Work** Web App. \nThe API is based on the following core tech:\n- NodeJS V12+ \n- ExpressJS V4+\n- MongoDB V4+\n",
    "version": "1.0.0",
    "title": "Ideal Place To Work - API"
  },
  "host": "localhost:9000",
  "basePath": "/api/v1/",
...
  "schemes": ["http"],
  "paths": {
...
    "/user/:userId": {
      "patch": {
        "tags": ["User"],
        "summary": "Updated user",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateUser",
        "produces": ["application/xml", "application/json"],
        "parameters": [
          {
            "name": "userId",
            "in": "url",
            "description": "userId that need to be updated",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid user supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": ["User"],
        "summary": "Delete user",
        "description": "This can only be done by the logged in user.",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "userId",
            "in": "url",
            "description": "The iuserId that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid userId supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "user_auth": {
      "type": "apiKey",
      "authorizationUrl": "http://lcalhost:9000/api/v1/login",
      "flow": "implicit",
      "scopes": {
        "write:users": "modify an user",
        "read:users": "read user"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "required": [
        "username",
        "firstname",
        "lastname",
        "createdAt",
        "updatedAt"
      ],
      "properties": {
        "username": { "type": "string" },
        "picture": { "type": "string", "required": false },
        "firstname": { "type": "string" },
        "lastname": { "type": "string" },
        "favouritePlaces": { "type": "array", "items": [] },
        "addedPlaces": { "type": "array", "items": [] },
        "createdAt": { "type": "string", "format": "date-time" },
        "updatedAt": { "type": "string", "format": "date-time" },
        "_id": { "type": "string" },
        "hash": { "type": "string" },
        "salt": { "type": "string" }
      },
      "xml": {
        "name": "User"
      }
    }
  },
...
}